refactor: 캐릭터 상수 public api 정리 (#119)
